move queries to db crate
